# Rewrite in C++ to improve performance