Use std::string, std::vector and std::map instead of T_STRING, T_ARRAY and T_HASH .

## Before

```
Calculating -------------------------------------
                Haml   714.000  i/100ms
        Faml (Array)     2.644k i/100ms
       Faml (String)     2.380k i/100ms
      Hamlit (Array)     1.422k i/100ms
     Hamlit (String)     1.306k i/100ms
        Slim (Array)     2.017k i/100ms
       Slim (String)     1.603k i/100ms
-------------------------------------------------
                Haml      8.017k (± 1.3%) i/s -     40.698k
        Faml (Array)     39.366k (± 3.0%) i/s -    198.300k
       Faml (String)     34.409k (± 3.4%) i/s -    173.740k
      Hamlit (Array)     18.288k (± 2.5%) i/s -     92.430k
     Hamlit (String)     17.162k (± 2.6%) i/s -     86.196k
        Slim (Array)     28.909k (± 3.1%) i/s -    145.224k
       Slim (String)     20.886k (± 2.3%) i/s -    105.798k

Comparison:
        Faml (Array):    39365.6 i/s
       Faml (String):    34408.6 i/s - 1.14x slower
        Slim (Array):    28909.3 i/s - 1.36x slower
       Slim (String):    20886.1 i/s - 1.88x slower
      Hamlit (Array):    18288.4 i/s - 2.15x slower
     Hamlit (String):    17162.3 i/s - 2.29x slower
                Haml:     8016.8 i/s - 4.91x slower
```

## After

```
Calculating -------------------------------------
                Haml   749.000  i/100ms
        Faml (Array)     3.327k i/100ms
       Faml (String)     2.794k i/100ms
      Hamlit (Array)     1.489k i/100ms
     Hamlit (String)     1.359k i/100ms
        Slim (Array)     2.084k i/100ms
       Slim (String)     1.595k i/100ms
-------------------------------------------------
                Haml      8.416k (± 1.2%) i/s -     42.693k
        Faml (Array)     53.435k (± 2.3%) i/s -    269.487k
       Faml (String)     44.437k (± 2.3%) i/s -    223.520k
      Hamlit (Array)     18.639k (± 1.6%) i/s -     93.807k
     Hamlit (String)     17.285k (± 1.7%) i/s -     86.976k
        Slim (Array)     28.898k (± 2.5%) i/s -    145.880k
       Slim (String)     21.367k (± 2.2%) i/s -    106.865k

Comparison:
        Faml (Array):    53435.2 i/s
       Faml (String):    44437.5 i/s - 1.20x slower
        Slim (Array):    28898.1 i/s - 1.85x slower
       Slim (String):    21366.7 i/s - 2.50x slower
      Hamlit (Array):    18639.3 i/s - 2.87x slower
     Hamlit (String):    17285.2 i/s - 3.09x slower
                Haml:     8416.3 i/s - 6.35x slower
```